Validates filename and node count
