+++ requirements.txt
@@ -7,3 +7,4 @@ simulaqron==3.0.15
 eqsn==0.0.8
 # qutip==4.5.2
 scipy==1.5.4
+pathvalidate==2.4.1
